Add createLightTheme and createDarkTheme
