+++ src/themes/index.js
@@ -9,12 +9,15 @@ import {DarkTheme} from './dark-theme/dark-theme.js';
 import {DarkThemeMove} from './move-theme/dark-theme-with-move.js';
 import {LightTheme} from './light-theme/light-theme.js';
 import {LightThemeMove} from './move-theme/light-theme-with-move.js';
-import createTheme from './create-theme.js';
+import createDarkTheme from './light-theme/create-dark-theme.js';
+import createLightTheme from './light-theme/create-light-theme.js';
 import darkThemePrimitives from './dark-theme/primitives.js';
 import lightThemePrimitives from './light-theme/primitives.js';
 
 export {
-  createTheme,
+  createDarkTheme,
+  createLightTheme,
+  createLightTheme as createTheme,
   DarkTheme as darkThemeOverrides,
   DarkTheme,
   DarkThemeMove,
